# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Bootcamp Meli - Desafio quality"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Property",
      "description": "Everything about properties",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/property": {
      "post": {
        "tags": [
          "Property"
        ],
        "summary": "Add a new property",
        "description": "",
        "operationId": "addPet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Property object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Property"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/property/{propertyId}": {
      "get": {
        "tags": [
          "Property"
        ],
        "summary": "Find property by ID",
        "description": "Returns a single property",
        "operationId": "getPetById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "description": "ID of property to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/PropertyFound"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Room": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "width": {
          "type": "number",
          "format": "int64"
        },
        "length": {
          "type": "number",
          "format": "int64"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Property": {
      "type": "object",
      "required": [
        "propName"
      ],
      "properties": {
        "propName": {
          "type": "string",
          "example": "Property name"
        },
        "propDistrict": {
          "type": "string"
        },
        "valueDistrictM2": {
          "type": "string"
        },
        "rooms": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/Room"
          }
        }
      }
    },
    "PropertyFound": {
      "type": "object",
      "required": [
        "propName"
      ],
      "properties": {
        "propName": {
          "type": "string",
          "example": "Property name"
        },
        "propDistrict": {
          "type": "string"
        },
        "totalPrice": {
          "type": "number"
        },
        "totalArea": {
          "type": "number"
        },
        "rooms": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Room"
          }
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}